requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
PyMySQL==1.1.1
SQLAlchemy==2.0.38
pydantic==2.10.6
python-dotenv==1.0.1
pandas==2.2.3